minor tweaks - still buggy

+++ Settings.cs
@@ -15,7 +15,7 @@ public static class Settings
     public static int HalfHeight = Height / 2;
     public static int TargetFps = 60;
     public static int FOV = 90;
-    public static int HalfFOV = FOV / 2;
+    public static float HalfFOV = FOV / 2f;
     public static float PlayerSpeed = 0.3f;
     public static float PlayerRotationSpeed = 0.12f;
     public static int PlayerHeight = 41;
